Parse mms pdu headers in 2 stages
MMS messages encode headers in a way in which it is possible to tell how long
the header data is without actually parsing that header data. This takes
advantage of that to split headers into a header name and the bytes of header
data, then loop over those headers and parse the header data later.